+++ Examples/MagicWeatherSwiftUI/Shared/Constants.swift
@@ -17,7 +17,7 @@ struct Constants {
      The API key for your app from the RevenueCat dashboard: https://app.revenuecat.com
      */
     #error("Modify this property to reflect your app's API key, then comment this line out.")
-    static let apiKey = "api_key"
+    static let apiKey = "REVENUECAT_API_KEY"
     
     /*
      The entitlement ID from the RevenueCat dashboard that is activated upon successful in-app purchase for the duration of the purchase.