+++ Sources/Purchasing/Purchases/Purchases.swift
@@ -342,6 +342,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
                     currentUserProvider: identityManager,
                     transactionsManager: transactionsManager,
                     deviceCache: deviceCache,
+                    offeringsManager: offeringsManager,
                     manageSubscriptionsHelper: manageSubsHelper,
                     beginRefundRequestHelper: beginRefundRequestHelper,
                     storeKit2TransactionListener: StoreKit2TransactionListener(delegate: nil),
@@ -360,6 +361,7 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
                     currentUserProvider: identityManager,
                     transactionsManager: transactionsManager,
                     deviceCache: deviceCache,
+                    offeringsManager: offeringsManager,
                     manageSubscriptionsHelper: manageSubsHelper,
                     beginRefundRequestHelper: beginRefundRequestHelper
                 )
@@ -457,6 +459,8 @@ public typealias StartPurchaseBlock = (@escaping PurchaseCompletedBlock) -> Void
 
         if self.systemInfo.dangerousSettings.autoSyncPurchases {
             storeKitWrapper.delegate = purchasesOrchestrator
+        } else {
+            Logger.warn(Strings.configure.autoSyncPurchasesDisabled)
         }
 
         self.subscribeToAppStateNotifications()
@@ -748,11 +752,10 @@ public extension Purchases {
         return self.customerInfoManager.customerInfoStream
     }
 
-    // swiftlint:disable line_length
     /**
      * Fetches the ``StoreProduct``s for your IAPs for given `productIdentifiers`.
      *
-     * Use this method if you aren't using ``getOfferings(completion:)``.
+     * Use this method if you aren't using ``Purchases/getOfferings(completion:)``.
      * You should use ``getOfferings(completion:)`` though.
      *
      * - Note: `completion` may be called without ``StoreProduct``s that you are expecting. This is usually caused by
@@ -760,7 +763,7 @@ public extension Purchases {
      * Also ensure that you have an active developer program subscription and you have signed the latest paid
      * application agreements.
      * If you're having trouble, see:
-     *  [App Store Connect In-App Purchase Configuration](https://www.revenuecat.com/2018/10/11/configuring-in-app-products-is-hard)
+     *  [App Store Connect In-App Purchase Configuration](https://rev.cat/how-to-configure-products)
      *
      * - Parameter productIdentifiers: A set of product identifiers for in-app purchases setup via
      * [AppStoreConnect](https://appstoreconnect.apple.com/)
@@ -773,9 +776,7 @@ public extension Purchases {
     func getProducts(_ productIdentifiers: [String], completion: @escaping ([StoreProduct]) -> Void) {
         purchasesOrchestrator.products(withIdentifiers: productIdentifiers, completion: completion)
     }
-    // swiftlint:enable line_length
 
-    // swiftlint:disable line_length
     /**
      * Fetches the ``StoreProduct``s for your IAPs for given `productIdentifiers`.
      *
@@ -787,7 +788,7 @@ public extension Purchases {
      * Also ensure that you have an active developer program subscription and you have signed the latest paid
      * application agreements.
      * If you're having trouble, see:
-     * [App Store Connect In-App Purchase Configuration](https://www.revenuecat.com/2018/10/11/configuring-in-app-products-is-hard)
+     * [App Store Connect In-App Purchase Configuration](https://rev.cat/how-to-configure-products)
      *
      * - Parameter productIdentifiers: A set of product identifiers for in-app purchases setup via
      * [AppStoreConnect](https://appstoreconnect.apple.com/)
@@ -798,7 +799,6 @@ public extension Purchases {
     func products(_ productIdentifiers: [String]) async -> [StoreProduct] {
         return await productsAsync(productIdentifiers)
     }
-    // swiftlint:enable line_length
 
     /**
      * Initiates a purchase of a ``StoreProduct``.
@@ -1015,7 +1015,7 @@ public extension Purchases {
      *
      * - Note: This method will not trigger a login prompt from App Store. However, if the receipt currently
      * on the device does not contain subscriptions, but the user has made subscription purchases, this method
-     * won't be able to restore them. Use ``restorePurchases(completion:)`` to cover those cases.
+     * won't be able to restore them. Use ``Purchases/restorePurchases(completion:)`` to cover those cases.
      */
     @objc func syncPurchases(completion: ((CustomerInfo?, Error?) -> Void)?) {
         purchasesOrchestrator.syncPurchases {
@@ -1035,7 +1035,7 @@ public extension Purchases {
      *
      * - Note: This method will not trigger a login prompt from App Store. However, if the receipt currently
      * on the device does not contain subscriptions, but the user has made subscription purchases, this method
-     * won't be able to restore them. Use ``restorePurchases(completion:)`` to cover those cases.
+     * won't be able to restore them. Use ``Purchases/restorePurchases(completion:)`` to cover those cases.
      */
     @available(iOS 13.0, macOS 10.15, tvOS 13.0, watchOS 6.2, *)
     func syncPurchases() async throws -> CustomerInfo {
@@ -1436,6 +1436,7 @@ public extension Purchases {
                   appUserID: configuration.appUserID,
                   observerMode: configuration.observerMode,
                   userDefaults: configuration.userDefaults,
+                  platformInfo: configuration.platformInfo,
                   storeKit2Setting: configuration.storeKit2Setting,
                   storeKitTimeout: configuration.storeKit1Timeout,
                   networkTimeout: configuration.networkTimeout,
@@ -1495,6 +1496,7 @@ public extension Purchases {
                                                       appUserID: String?,
                                                       observerMode: Bool,
                                                       userDefaults: UserDefaults?,
+                                                      platformInfo: PlatformInfo?,
                                                       storeKit2Setting: StoreKit2Setting,
                                                       storeKitTimeout: TimeInterval,
                                                       networkTimeout: TimeInterval,
@@ -1503,7 +1505,7 @@ public extension Purchases {
                                   appUserID: appUserID,
                                   userDefaults: userDefaults,
                                   observerMode: observerMode,
-                                  platformInfo: nil,
+                                  platformInfo: platformInfo,
                                   storeKit2Setting: storeKit2Setting,
                                   storeKitTimeout: storeKitTimeout,
                                   networkTimeout: networkTimeout,