+++ Sources/Purchasing/StoreKit2/ProductsFetcherSK2.swift
@@ -29,10 +29,20 @@ actor ProductsFetcherSK2 {
     func products(identifiers: Set<String>) async throws -> Set<SK2StoreProduct> {
         do {
             if let cachedProducts = await self.cachedProducts(withIdentifiers: identifiers) {
+                Logger.debug(
+                    Strings.offering.products_already_cached(
+                        identifiers: Set(cachedProducts.map { $0.productIdentifier})
+                    )
+                )
                 return cachedProducts
             }
 
+            Logger.debug(
+                Strings.storeKit.no_cached_products_starting_store_products_request(identifiers: identifiers)
+            )
+
             let storeKitProducts = try await StoreKit.Product.products(for: identifiers)
+            Logger.rcSuccess(Strings.storeKit.store_product_request_received_response)
             let sk2StoreProducts = Set(storeKitProducts.map { SK2StoreProduct(sk2Product: $0) })
 
             await self.cache(products: sk2StoreProducts)
@@ -43,8 +53,16 @@ actor ProductsFetcherSK2 {
         }
     }
 
-    func clearCache() {
-        self.cachedProductsByIdentifier.removeAll(keepingCapacity: false)
+    /// - Returns: The product identifiers that were removed, or empty if there were not
+    ///   cached products.
+    @discardableResult
+    func clearCache() -> Set<String> {
+        let cachedProductIdentifiers = self.cachedProductsByIdentifier.keys
+        if !cachedProductIdentifiers.isEmpty {
+            Logger.debug(Strings.offering.product_cache_invalid_for_storefront_change)
+            self.cachedProductsByIdentifier.removeAll(keepingCapacity: false)
+        }
+        return Set(cachedProductIdentifiers)
     }
 
 }