+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## 4.9.1
+### Fixes:
+* `CustomerInfoResponseHandler`: return `CustomerInfo` instead of error if the response was successful (#1778) via NachoSoto (@NachoSoto)
+* Error logging: `logErrorIfNeeded` no longer prints message if it's the same as the error description (#1776) via NachoSoto (@NachoSoto)
+* fix another broken link in docC docs (#1777) via aboedo (@aboedo)
+* fix links to restorePurchase (#1775) via aboedo (@aboedo)
+* fix getProducts docs broken link (#1772) via aboedo (@aboedo)
+
+### Improvements:
+* `Logger`: wrap `message` in `@autoclosure` to avoid creating when `LogLevel` is disabled (#1781) via NachoSoto (@NachoSoto)
+
+### Other changes:
+* Lint: fixed `SubscriberAttributesManager` (#1774) via NachoSoto (@NachoSoto)
+## 4.9.0
+* Update Configuration.swift to include platformInfo. Used by PurchasesHybridCommon (#1760) via Joshua Liebowitz (@taquitos)
+
+## 4.8.0
+### New API
+
+* `EntitlementInfo`: added `isActiveInCurrentEnvironment` and `isActiveInAnyEnvironment` (#1755) via NachoSoto (@NachoSoto)
+
+### Other Changes
+* Plumb platformInfo in Configuration for PHC use (#1757) via Joshua Liebowitz (@taquitos)
+* added a log when `autoSyncPurchases` is disabled (#1749) via aboedo (@aboedo)
+* Re-fetch cached offerings and products after Storefront changes (3/4)  (#1743) via Juanpe Catalán (@Juanpe)
+* `bug_report.md`: clarify SK2 support (#1752) via NachoSoto (@NachoSoto)
+* `logErrorIfNeeded`: also log message if present (#1754) via NachoSoto (@NachoSoto)
+
 ## 4.7.0
 ### Changes:
 * Replaced `CustomerInfo.nonSubscriptionTransactions` with a new non-`StoreTransaction` type (#1733) via NachoSoto (@NachoSoto)